requirements.txt
asgiref==3.6.0
Django==3.2.16
django-environ==0.9.0
gunicorn==20.1.0
pytz==2022.7.1
sqlparse==0.4.3
django-filter==21.1
djangorestframework-simplejwt==5.2.2
PyMySQL==1.0.2
djangorestframework==3.12.4
django-cors-headers
